+++ prereqs_installer.sh
@@ -3,7 +3,7 @@
 # Install all of the packages and dependencies required for NadekoBot to run on Linux
 # distributions.
 #
-# Comment key for '[letter].[number].':
+# Comment key:
 #   A.1. - NOTE: If the write perms are not applied to all users for this tool, attempts
 #                to update 'youtube-dl' by a non-root user will always fail.
 #   B.1. - FIXME: Find a better solution than modifying the perms in such a way that I
@@ -47,8 +47,7 @@ install_prereqs() {
 
     sudo curl -s -L https://yt-dl.org/downloads/latest/youtube-dl -o \
         /usr/local/bin/youtube-dl
-    # A.1.
-    # B.1.
+    # A.1. & B.1.
     sudo chmod a+rwx /usr/local/bin/youtube-dl
 }
 
@@ -78,10 +77,8 @@ read -rp "We will now install NadekoBot's prerequisites. Press [Enter] to contin
 #   20.04
 if [[ $_DISTRO = "ubuntu" ]]; then
     case "$_VER" in
-        16.04) install_prereqs "ubuntu" "16.04" "python" ;;
-        18.04) install_prereqs "ubuntu" "18.04" "python" ;;
-        20.04) install_prereqs "ubuntu" "20.04" "python-is-python3" ;;
-        *)     unsupported ;;
+        16.04|18.04|20.04) install_prereqs "ubuntu" "$_VER" "python" ;;
+        *)                 unsupported ;;
     esac
 # Linux Mint:
 #   9
@@ -116,8 +113,7 @@ elif [[ $_DISTRO = "debian" ]]; then
 
     sudo curl -s -L https://yt-dl.org/downloads/latest/youtube-dl -o \
         /usr/local/bin/youtube-dl
-    # A.1.
-    # B.1.
+    # A.1. & B.1.
     sudo chmod a+rwx /usr/local/bin/youtube-dl
             ;;
         10) install_prereqs "debian" "10" "python" ;;